feat(ehr): fixes
Ref: #1040
